BUG in ``log_mask_zero``: log(0) = -inf (#217)
* BUG in ``log_mask_zero``: log(0) = -inf

Commit f66a053e7f6b1081f000cd871b73b04ea4a62433 introduced masking of
probabilities that are zero to avoid run-time warning of their logarithm.
It was assumed that the log(x) is 0 for x <= 0. But this corresponds to
x = 1. The right value is -inf for x = 0.

This fix sets log(0) = -inf. Additionally, all values of log(x) for
x < 0 are also set to -inf, which is reasonable considering that x is a
probability and should not be smaller than 0 in the first place.

* ENH preserve NaNs in ``log_mask_zero``

And the respective warning for values below zero as proposed by
@vrenkens in https://github.com/hmmlearn/hmmlearn/pull/217#issuecomment-327468000.